Complete CRUD impl, separate files, fix clippy
